requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
slack-sdk==3.4.2
urllib3==1.26.4
Werkzeug==1.0.1
flask_httpauth